Moved BulkUpload actions from OdsInstanceSettingsController to a separate BulkUploadController. Refactored to remove the unused components in OdsInstanceSettingsController.
