+++ Application/EdFi.Ods.AdminApp.Web/Controllers/OdsInstanceSettingsController.cs
@@ -3,49 +3,31 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-using System;
-using System.IO;
-using System.Linq;
-using System.Net;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Hosting;
 using EdFi.Ods.AdminApp.Management;
 using EdFi.Ods.AdminApp.Management.Api;
 using EdFi.Ods.AdminApp.Management.Helpers;
-using EdFi.Ods.AdminApp.Management.Instances;
 using EdFi.Ods.AdminApp.Management.Settings;
 using EdFi.Ods.AdminApp.Web.Display.TabEnumeration;
 using EdFi.Ods.AdminApp.Web.Infrastructure;
-using EdFi.Ods.AdminApp.Web.Infrastructure.IO;
 using EdFi.Ods.AdminApp.Web.Infrastructure.Jobs;
 using EdFi.Ods.AdminApp.Web.Models.ViewModels;
 using EdFi.Ods.AdminApp.Web.Models.ViewModels.OdsInstanceSettings;
-using FluentValidation;
-using log4net;
 using Microsoft.Extensions.Options;
 
 namespace EdFi.Ods.AdminApp.Web.Controllers
 {
     public class OdsInstanceSettingsController : ControllerBase
     {
-        private readonly IOdsApiConnectionInformationProvider _apiConnectionInformationProvider;
         private readonly ICachedItems _cachedItems;
         private readonly ICloudOdsSettingsService _cloudOdsSettingsService;
         private readonly IGetProductionApiProvisioningWarningsQuery _getProductionApiProvisioningWarningsQuery;
-        private readonly ILog _logger = LogManager.GetLogger(typeof(OdsInstanceSettingsController));
         private readonly IOdsApiFacadeFactory _odsApiFacadeFactory;
-        private readonly IOdsSecretConfigurationProvider _odsSecretConfigurationProvider;
         private readonly IProductionSetupJob _productionSetupJob;
         private readonly ITabDisplayService _tabDisplayService;
-        private readonly string _missingOdsSecretConfig = "ODS secret configuration can not be null";
-        private readonly IFileUploadHandler _fileUploadHandler;
-        private readonly IBulkUploadJob _bulkUploadJob;
         private readonly InstanceContext _instanceContext;
-        private readonly IInferOdsApiVersion _inferOdsApiVersion;   
-        private readonly IValidator<BulkFileUploadModel> _bulkLoadValidator;
         private readonly AppSettings _appSettings;
-        private readonly IWebHostEnvironment _webHostEnvironment;
 
         public OdsInstanceSettingsController(
               IOdsApiFacadeFactory odsApiFacadeFactory
@@ -54,15 +36,8 @@ namespace EdFi.Ods.AdminApp.Web.Controllers
             , IProductionSetupJob productionSetupJob
             , ICloudOdsSettingsService cloudOdsSettingsService
             , ITabDisplayService tabDisplayService 
-            , IOdsSecretConfigurationProvider odsSecretConfigurationProvider
-            , IOdsApiConnectionInformationProvider apiConnectionInformationProvider
-            , IFileUploadHandler fileUploadHandler
-            , IBulkUploadJob bulkUploadJob
             , InstanceContext instanceContext
-            , IInferOdsApiVersion inferOdsApiVersion          
-            , IValidator<BulkFileUploadModel> bulkLoadValidator
             , IOptions<AppSettings> appSettingsAccessor
-            , IWebHostEnvironment webHostEnvironment
             )
         {
             _odsApiFacadeFactory = odsApiFacadeFactory;
@@ -71,15 +46,8 @@ namespace EdFi.Ods.AdminApp.Web.Controllers
             _productionSetupJob = productionSetupJob;
             _cloudOdsSettingsService = cloudOdsSettingsService;
             _tabDisplayService = tabDisplayService;
-            _odsSecretConfigurationProvider = odsSecretConfigurationProvider;
-            _apiConnectionInformationProvider = apiConnectionInformationProvider;
-            _fileUploadHandler = fileUploadHandler;
-            _bulkUploadJob = bulkUploadJob;
             _instanceContext = instanceContext;
-            _inferOdsApiVersion = inferOdsApiVersion;
-            _bulkLoadValidator = bulkLoadValidator;
             _appSettings = appSettingsAccessor.Value;
-            _webHostEnvironment = webHostEnvironment;
         }
 
         public async Task<ActionResult> Setup()
@@ -139,137 +107,6 @@ namespace EdFi.Ods.AdminApp.Web.Controllers
             };
 
             return View(model);
-        }
-
-        private async Task<OdsApiConnectionInformation> GetConnectionInformationProvider()
-        {
-            return await _apiConnectionInformationProvider.GetConnectionInformationForEnvironment();
-        }
-
-        public async Task<ActionResult> BulkLoad()
-        {
-            var config = await _odsSecretConfigurationProvider.GetSecretConfiguration(_instanceContext.Id);
-            if (config == null)
-            {
-                _logger.Error(_missingOdsSecretConfig);
-                return JsonError(_missingOdsSecretConfig);
-            }
-            var model = new OdsInstanceSettingsModel
-            {
-                OdsInstanceSettingsTabEnumerations =
-                    _tabDisplayService.GetOdsInstanceSettingsTabDisplay(OdsInstanceSettingsTabEnumeration
-                        .BulkLoad),
-                BulkFileUploadModel = new BulkFileUploadModel
-                {
-                    ApiKey = config.BulkUploadCredential?.ApiKey ?? string.Empty,
-                    ApiSecret = config.BulkUploadCredential?.ApiSecret ?? string.Empty
-                },
-                OdsInstance = _instanceContext
-            };
-            return View(model);
-        }
-
-        [HttpPost]
-        public async Task<ActionResult> BulkFileUpload(OdsInstanceSettingsModel model)
-        {
-            var bulkFiles = model.BulkFileUploadModel.BulkFiles.Where(file => file != null && file.Length > 0).ToArray();
-
-            if (!bulkFiles.Any())
-            {
-                return NoContent();
-            }
-
-            if (bulkFiles.Sum(f => f.Length) > BulkFileUploadModel.MaxFileSize)
-            {
-                throw new Exception($"Upload exceeds maximum limit of {BulkFileUploadModel.MaxFileSize} bytes");
-            }
-
-            if (bulkFiles.Length > 1)
-            {
-                throw new Exception("Currently, the bulk import process only supports a single file at a time");
-            }
-
-            var uploadedFiles = _fileUploadHandler.SaveFilesToUploadDirectory(bulkFiles,
-                fileName => InterchangeFileHelpers.BuildFileNameForImport(model.BulkFileUploadModel.BulkFileType, fileName));
-
-            var connectionInformation = await GetConnectionInformationProvider();
-
-            var config = await _odsSecretConfigurationProvider.GetSecretConfiguration(_instanceContext.Id);
-            if (config == null)
-            {
-                throw new InvalidOperationException("ODS secret configuration can not be null.");
-            }
-
-            if (model.BulkFileUploadModel != null)
-            {               
-                model.BulkFileUploadModel.ApiKey = config.BulkUploadCredential?.ApiKey;
-                var validationResult = await _bulkLoadValidator.ValidateAsync(model.BulkFileUploadModel);
-                if (!validationResult.IsValid)
-                {
-                    var errorMessage = string.Join(",", validationResult.Errors.Select(x => x.ErrorMessage));
-                    Response.StatusCode = (int) HttpStatusCode.BadRequest;
-                    return Json(new {Result = new {Errors = new[] {new {ErrorMessage = errorMessage}}}});
-                }
-            }
-
-            var schemaBasePath = Path.Combine(_webHostEnvironment.ContentRootPath, _appSettings.XsdFolder);
-            var standardVersion = _inferOdsApiVersion.EdFiStandardVersion(connectionInformation.ApiServerUrl);
-            var odsApiVersion = _inferOdsApiVersion.Version(connectionInformation.ApiServerUrl);
-
-            const int IdealSimultaneousRequests = 20;
-            const int PessimisticSimultaneousRequests = 1;
-
-            var maxSimultaneousRequests = odsApiVersion.StartsWith("3.")
-                ? PessimisticSimultaneousRequests
-                : IdealSimultaneousRequests;
-
-            var jobContext = new BulkUploadJobContext
-            {
-                DataDirectoryFullPath = uploadedFiles.Directory,
-                OdsInstanceId = _instanceContext.Id,
-                OdsInstanceName = _instanceContext.Name,
-                ApiBaseUrl = connectionInformation.ApiBaseUrl,
-                OauthUrl = connectionInformation.OAuthUrl,
-                MetadataUrl = connectionInformation.MetadataUrl,
-                DependenciesUrl = connectionInformation.DependenciesUrl,
-                ClientKey = config.BulkUploadCredential?.ApiKey ?? string.Empty,
-                ClientSecret = config.BulkUploadCredential?.ApiSecret ?? string.Empty,
-                SchemaPath = $"{schemaBasePath}\\{standardVersion}",
-                MaxSimultaneousRequests = maxSimultaneousRequests
-            };
-
-            if (!_bulkUploadJob.IsJobRunning())
-            {
-                _bulkUploadJob.EnqueueJob(jobContext);
-            }
-
-            var updatedValue = model;
-            updatedValue.BulkFileUploadModel.IsJobRunning = _bulkUploadJob.IsJobRunning();
-            updatedValue.BulkFileUploadModel.IsSameOdsInstance = _bulkUploadJob.IsSameOdsInstance(_instanceContext.Id, typeof(BulkUploadJobContext));
-            return PartialView("_SignalRStatus_BulkLoad", updatedValue);
-        }
-
-        [HttpPost]
-        public async Task<ActionResult> SaveBulkLoadCredentials(SaveBulkUploadCredentialsModel model)
-        {
-            var config = await _odsSecretConfigurationProvider.GetSecretConfiguration(_instanceContext.Id);
-            if (config == null) return JsonError(_missingOdsSecretConfig);
-            config.BulkUploadCredential = new BulkUploadCredential { ApiKey = model.ApiKey,
-                ApiSecret = model.ApiSecret };
-            await _odsSecretConfigurationProvider.SetSecretConfiguration(config, _instanceContext.Id);
-            return JsonSuccess("Credentials successfully saved");
-        }
-
-        [HttpPost]
-        public async Task<ActionResult> ResetCredentials()
-        {
-            var config = await _odsSecretConfigurationProvider.GetSecretConfiguration(_instanceContext.Id);
-            if (config == null) return JsonError(_missingOdsSecretConfig);
-            if (config.BulkUploadCredential == null) return JsonError("Missing bulk load credentials");
-            config.BulkUploadCredential.ApiKey = string.Empty;
-            config.BulkUploadCredential.ApiSecret = string.Empty;
-            await _odsSecretConfigurationProvider.SetSecretConfiguration(config, _instanceContext.Id);
-            return JsonSuccess("Credentials successfully reset");
         }
     }
 }
